+++ tests/run-base.R
@@ -1,3 +1,7 @@
 library(testthat)
-set.seed(getOption("mlr.debug.seed"))
-test_check("mlr", filter = "base_")
+
+# no tests on CRAN
+if (identical(Sys.getenv("NOT_CRAN"), "true")) {
+  set.seed(getOption("mlr.debug.seed"))
+  test_check("mlr", filter = "base_")
+}