+++ tests/run-learners-cluster.R
@@ -1,4 +1,8 @@
 library(testthat)
-set.seed(getOption("mlr.debug.seed"))
 
-test_check("mlr", "_learners_all_cluster")
+# no tests on CRAN
+if (identical(Sys.getenv("NOT_CRAN"), "true")) {
+
+  set.seed(getOption("mlr.debug.seed"))
+  test_check("mlr", "_learners_all_cluster")
+}