+++ tests/run-multilabel.R
@@ -1,4 +1,8 @@
 library(testthat)
-set.seed(getOption("mlr.debug.seed"))
 
-test_check("mlr", "_multilabel_")
+# no tests on CRAN
+if (identical(Sys.getenv("NOT_CRAN"), "true")) {
+
+  set.seed(getOption("mlr.debug.seed"))
+  test_check("mlr", "_multilabel_")
+}